Squashed changes for refactor of SyntheticVolumeBase class structure.
